+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - radio button component added
 - avatar component added
 - alert component added
+- switch component added
 - simple table component added
 - breadcrumb component changed props
 - drawer component now work with v-model and $drawer() not supported